+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v1.2.1
+#### 🐛 Fixes
+
+- Kill server after Unity crash (PR: #101)
+- Persist chat template on remote servers (PR: #103)
+
+
 ## v1.2.0
 #### 🚀 Features